+++ components/sara-chat/sara-chat-list.tsx
@@ -3,18 +3,22 @@
 import { SaraSession } from 'auth'
 import SaraLoading from 'components/sara-loading'
 import { IconRefresh } from 'components/ui/icons'
-import { isPreviewFeatureEnabled } from 'lib/service-utils'
 
 import { ChatQuery } from './../../lib/data-model-types'
 import { Separator } from './../ui/separator'
 import SaraChatQueryContent from './sara-chat-query-content'
+import { updateResource } from 'app/saraClient'
 
 export interface SaraChatListProps {
+  chatId: string | undefined
   chatQueries: ChatQuery[]
   saraSession: SaraSession
 }
 
-const SaraChatList = ({ chatQueries, saraSession }: SaraChatListProps) => {
+const buildChatQueryUrl = (chatId: string, chatQueryId: string) =>
+  `/api/chats/${chatId}/chat-queries/${chatQueryId}`
+
+const SaraChatList = ({ chatId, chatQueries, saraSession }: SaraChatListProps) => {
   if (!chatQueries.length) {
     return null
   }
@@ -56,10 +60,26 @@ const SaraChatList = ({ chatQueries, saraSession }: SaraChatListProps) => {
                       : undefined
                   }
                 />
-                {isPreviewFeatureEnabled('RegenerateResponse') && (
-                  <button className="flex items-center p-2 border-2 border-invisible hover:border-black text-sm leading-5 font-medium rounded-md text-black mb-2">
-                    <IconRefresh className="mr-2" />
-                    <p>Regenerate response</p>
+                {(index == chatQueries.length - 1) && (
+                  <button 
+                    className="flex items-center p-2 border-2 border-invisible hover:border-black text-sm leading-5 font-medium rounded-md text-black mb-2"
+                    onClick={async (e) => {
+                      const chatQueryUrl = buildChatQueryUrl(chatId!, chatQuery.id)
+                      const patchReqBody = {
+                        status: 'QUERY_RECEIVED',
+                      }
+
+                      const resubmittedChatQuery = await updateResource<ChatQuery>(
+                        chatQueryUrl,
+                        patchReqBody,
+                        `Failed to re-submit chat query '${chatQuery.id}' when in an 'ERROR' state`,
+                      )
+
+                      chatQueries[chatQueries.length - 1] = resubmittedChatQuery
+                    }}
+                    >
+                      <IconRefresh className="mr-2" />
+                      <p>Regenerate response</p>
                   </button>
                 )}
               </div>